+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic parsers: `any`, `str`, `re`, `or`, `seq`, `many`, `many1`, `maybe`,
   `not`, `tag`
 - `fail` creates a labelled failure, which is like a throw for parsers
+- `failLabel(parser, label)` is a shortcut for `or(parser, fail(label))`
 - `cstToAst` transforms a CST to an AST
 
 [unreleased]: https://github.com/unleashy/vahv/compare/v0.1.0...HEAD